No need to use bot IDS, use of the readable label instead
The strategies information can be found at https://middle.napbots.com/v1/strategy. It is in json format and contains the corresponding ID for each bot.
The new function get_strategies_code() parses the json file.
The function assign_composition() as  een adapted to 'translate' the configured compositions to coded compositions (aka with IDS). In this function "global $debug;" was missing.